# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.GithubMirror:
    properties:
      base_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  util.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: FRP 管理面板后端 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FRP 管理面板 API
  version: "1.0"
paths:
...
  /api/github-mirrors:
    get:
      consumes:
      - application/json
      description: 获取所有已配置的GitHub加速源列表
      produces:
      - application/json
      responses:
        "200":
          description: 加速源列表
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "500":
          description: 获取加速源列表失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取所有GitHub加速源
      tags:
      - GitHub加速源
    post:
      consumes:
      - application/json
      description: 创建新的GitHub加速源配置
      parameters:
      - description: 加速源信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.GithubMirror'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 创建加速源失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 创建GitHub加速源
      tags:
      - GitHub加速源
  /api/github-mirrors/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的GitHub加速源配置
      parameters:
      - description: 加速源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 删除加速源失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 删除GitHub加速源
      tags:
      - GitHub加速源
    get:
      consumes:
      - application/json
      description: 根据ID获取指定的GitHub加速源详情
      parameters:
      - description: 加速源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 加速源详情
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: 加速源不存在
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取GitHub加速源详情
      tags:
      - GitHub加速源
    put:
      consumes:
      - application/json
      description: 更新指定的GitHub加速源配置
      parameters:
      - description: 加速源ID
        in: path
        name: id
        required: true
        type: integer
      - description: 加速源信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.GithubMirror'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 更新加速源失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 更新GitHub加速源
      tags:
      - GitHub加速源
  /api/github-mirrors/{id}/set-default:
    post:
      consumes:
      - application/json
      description: 将指定的GitHub加速源设置为默认使用的加速源
      parameters:
      - description: 加速源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: 设置默认加速源失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 设置默认GitHub加速源
      tags:
      - GitHub加速源
...
  /api/monitor/overview:
    get:
      consumes:
      - application/json
      description: 获取系统监控概览数据，包括服务器数量、客户端数量、代理数量等统计信息
      produces:
      - application/json
      responses:
        "200":
          description: 监控概览数据
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: 获取监控概览失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取监控概览
      tags:
      - 监控
  /api/monitor/stats:
    get:
      consumes:
      - application/json
      description: 获取系统详细统计数据，包括流量统计、连接数�?
      produces:
      - application/json
      responses:
        "200":
          description: 统计数据
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: 获取统计数据失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取统计数据
      tags:
      - 监控
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"